实现命令模式

+++ iOSDesignPatternDemo/iOSDesignPatternDemo/API/LibraryAPI.h
@@ -16,7 +16,7 @@
 
 - (NSArray *)getAlbums;
 - (void)addAlbum:(Album *)album atIndex:(int)index;
-- (void)deleteAlbumAtIndex:(int)index;
+- (void)deleteAlbumAtIndex:(NSInteger)index;
 
 - (void)saveAlbums;
